refactor: split ExecutionWorkspace into colocated hooks and feature components (#304)
Break the ~1370-line ExecutionWorkspace orchestrator into feature-focused,
colocated files per the CodeRabbit plan, extending it to cover the agent-mode
section that landed with #285:

Hooks / pure utils:
- executionJobUtils: collectActivePassNames, resolveQueuedModelIdentifier,
  buildQueuedBatchJob, describeAppliedMcpPatches (+ unit tests)
- useDiagnosis: MCP diagnostic lifecycle, history, apply-fix mapping
- useOwrExport: OWR export overlay state + bundle download (fflate zip)
- useRecipeView: graph/json view state, More menu, export overlay actions

Presentational components:
- ExportRecipeOverlay, RecipePreviewCard, ExecutionLogPanel,
  ManualExecutionControls, AgentModeSection

ExecutionWorkspace.tsx remains a thin orchestrator wiring the hooks and
presentational sections. Behavior is unchanged: existing component tests pass
unmodified.
